great stable wip

+++ src/Channel_Signal.h
@@ -19,7 +19,7 @@ class Channel_Signal
             m_y_offset            = CHANNEL_SIGNAL_Y_OFFSET;
 
     uint32_t m_raw_values[LAB_OSCILLOSCOPE_NUMBER_OF_SAMPLES_PER_CHANNEL];
-    std::vector<double>           m_values;
+    std::vector<float>           m_values;
     std::vector<std::vector<int>> m_pixel_points;
     
     // for function generator
